% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlab.R
\name{alienNativeRichness}
\alias{alienNativeRichness}
\title{The function calculates alien and native species richness in a 
  
   @param aDatasetWith_eunishabitatstypename_alien_locality_eunisspeciesgroup_scientificname dataset containing
   the following columns-ranges (colnames are compliant with lifewatch dataportal harmonization):
   "eunishabitatstypename": (string) one of the eunishabitat codes like "C1.1" etc. 
     (See: http://eunis.eea.europa.eu/habitats-code-browser.jsp?expand=C#level_C)
     The current EEA linked data implementation seems not to dereference properly 
     eunishabitatstypenames URIs (of the form http://eunis.eea.europa.eu/eunishabitats/C1.1).
     The definition can be retrieved instead by the following service:
     http://semantic.eea.europa.eu/factsheet.action?uri=http://eunis.eea.europa.eu/eunishabitats/C1.1
     (passing the uri as the parameter value)
     or by the sparql query SELECT * WHERE{ <http://eunis.eea.europa.eu/eunishabitats/C1.1> ?p ?o. }
     running at the official eea endpoint:
     http://semantic.eea.europa.eu/sparql?selectedBookmarkName=&query=SELECT+*+WHERE%7B+%3Chttp%3A%2F%2Feunis.eea.europa.eu%2Feunishabitats%2FC1%3E+%3Fp+%3Fo.+%0D%0A+%7D&format=text%2Fhtml&nrOfHits=20&execute=Execute
     
   "alien": (boolean) an alien species with respect to the observing context 
   "locality": (string) a toponym where the observation occurred 
   [DEPRECATED] "eunisspeciesgroups": (string) species group according to eunis species list (cf. http://eunis.eea.europa.eu/species.jsp)
   "family": (string) family name according to LifeWatch Global Name Architecture - cf.(for auth users) http://www.servicecentrelifewatch.eu/global-names-architecture.
   "scientificname": (string) scientific name of the observed species (to be clarified: which harmonization we should expect)}
\usage{

  alienNativeRichness(aDatasetWith_eunishabitatstypename_alien_locality_eunisspeciesgroup_scientificname)
}
\value{
dataframe accounting for alien and native species richness. 
containing the fields: "locality", "EunisL1", "eunisspeciesgroup", "native_richness", "alien_richness"
original data are grouped by locality, eunis habitat level-1 code, eunis species group.
}
\description{
It results in a new dataset accounting for alien and native species richness, grouping the original data
by locality, eunis habitat level-1 code, family 
[deprecated] eunis species group.
}
\details{
NOTE: We want to investigate the site vulnerability, distributed across different Eunis habitat, to different taxonomic groups. 
So we need to aggregate the data at site level by taxon name and Habita Eunis name
}
\author{
Paolo Colangelo (original script author), Paolo Tagliolato (engineering)
}

