% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlab.R
\name{readLifeWatchDataportal}
\alias{readLifeWatchDataportal}
\title{Download a dataset from LifeWatch Service Centre dataportal}
\usage{
readLifeWatchDataportal(dataUID, user, pass)
}
\arguments{
\item{dataUID}{unique identifier of the dataset, retrievable through dataportal API or search interface}

\item{user}{username (LW dataportal account)}

\item{pass}{password (LW dataportal account)}
}
\value{
first row of the downloaded dataset
}
\description{
\code{readLifeWatchDataportal} The method returns the first row of the downloaded dataset
In openCPU, when the request is invoked with POST method, the dataset can then be found in /ocpu/tmp/{key}/files/lw_{dataUID}
The session {key} is given in "X-ocpu-session" http response header returned by openCPU. 
Alternatively, the complete path to session folder ({baseurl}/ocpu/tmp/{key}/) is given in "Location" header
}

